Flesh out Grid and Location
This is a finished first pass at work on a `Grid`.  It provides an
interface to create, modifify, and get information from.  A smaller
unit of data in the grid is represented by a `Location`; which has
the responsibiliby of describing what data is found at any given
coordinates of the `Grid`.